requirements.txt
napari-plugin-engine>=0.1.4
numpy>=0.1.19
zarr>=0.2.4
dask[array]>=2.20.0
openslide-python>=1.1.1
